requirements.txt
Django
djangorestframework
gTTS
Pillow
psycopg2-binary
pytesseract
gunicorn
drf-yasg
django-cors-headers
selenium
webdriver-manager